Add XcodeInspector
